+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   * Added support for embedding mojo.js applications with `mountPlugin`.
   * Added supprt for rewriting `ctx.req.path`.
+  * Added supprt relative paths with `ctx.req.basePath` to `ctx.urlFor`.
   * Added `app:warmup` application hook.
 
 ### Bug Fixes